[FIX] 사진 및 동영상 `Position` type 수정
map api 에서 가져오는 position 값이 소수점이기 때문에
Long  -> Double
로 변경

+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/CreatePostUiState.kt
@@ -11,6 +11,6 @@ sealed class PostBlock(open val content: String) {
 }
 
 data class Position(
-    val latitude: Long,
-    val longitude: Long
+    val latitude: Double,
+    val longitude: Double
 )
